btrfs: add support secondary hash
Authenticated hashes don't allow to do basic checksum verification
without the key. This leaves the users with all-or-nothing choice.

There's another option that weakens the authentication and stores
another (unauthenticated) hash in the available space.

This patch adds support for secondary hash that is hardcoded to be
XXHASH (64 bits). The 32 bytes are split to 24 + 8, which means
reduction from 256 to 192 bits for the authentication hash.

The idea of split hash has been mentioned in a LWN comment [1], ZFS uses
16 + 16 bytes of SHA512/SHA256.

[1] https://lwn.net/Articles/819143/

Signed-off-by: David Sterba <dsterba@suse.com>